21.04.26 Актулизация формул

+++ GKL Analise/GKL Analise/IPolotno.cs
@@ -2,6 +2,6 @@
 {
     internal interface IPolotno : IConstruction
     {
-        int VirezSq { get; }
+        double VirezSq { get; }
     }
 }
